Add functions comments and move interfaces to process
