+++ apps/vs-code-designer/src/app/commands/pickFuncProcess.ts
@@ -12,8 +12,7 @@ import { getProjFiles } from '../utils/dotnet/dotnet';
 import { isTimeoutError } from '../utils/requestUtils';
 import { executeIfNotActive } from '../utils/taskUtils';
 import { getWorkspaceSetting } from '../utils/vsCodeConfig/settings';
-import type { IProcessInfo, IWindowsProcessTree } from '../utils/windowsProcessTree';
-import { getWindowsProcessTree, ProcessDataFlag } from '../utils/windowsProcessTree';
+import { getWindowsProcessTree } from '../utils/windowsProcessTree';
 import type { HttpOperationResponse } from '@azure/ms-rest-js';
 import { delay } from '@azure/ms-rest-js';
 import { HTTP_METHODS } from '@microsoft/utils-logic-apps';
@@ -21,8 +20,8 @@ import type { AzExtRequestPrepareOptions } from '@microsoft/vscode-azext-azureut
 import { sendRequestWithTimeout } from '@microsoft/vscode-azext-azureutils';
 import { UserCancelledError } from '@microsoft/vscode-azext-utils';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
-import { ProjectLanguage } from '@microsoft/vscode-extension';
-import type { IPreDebugValidateResult } from '@microsoft/vscode-extension';
+import { ProcessDataFlag, ProjectLanguage } from '@microsoft/vscode-extension';
+import type { IPreDebugValidateResult, IProcessInfo, IWindowsProcessTree } from '@microsoft/vscode-extension';
 import * as unixPsTree from 'ps-tree';
 import * as vscode from 'vscode';
 import * as parse from 'yargs-parser';
