+++ libs/vscode-extension/src/lib/models/index.ts
@@ -13,3 +13,4 @@ export * from './templates';
 export * from './bundleFeed';
 export * from './tasks';
 export * from './launch';
+export * from './process';
